Refactor retrofit code in neopixel activity to new class as a service

+++ app/src/main/java/be/vives/android/nico/thumpercontrol/rest/neopixel/effects/NeoPixelStrobeEffect.java
@@ -7,13 +7,13 @@ public class NeoPixelStrobeEffect {
     private int red;
     private int green;
     private int blue;
-    private byte delay;
+    private int delay;
 
     public NeoPixelStrobeEffect() {
-        this(0, 0, 0, (byte)0);
+        this(0, 0, 0, 0);
     }
 
-    public NeoPixelStrobeEffect(int red, int green, int blue, byte delay) {
+    public NeoPixelStrobeEffect(int red, int green, int blue, int delay) {
         this.red = red;
         this.green = green;
         this.blue = blue;
